move isTokenExpired to utils
