+++ packages/core/src/utils/index.ts
@@ -16,3 +16,4 @@ export { urlSafeDecode } from './urlSafeDecode';
 export { urlSafeEncode } from './urlSafeEncode';
 export { deepFreeze } from './deepFreeze';
 export { deDupeAsyncFunction } from './deDupeAsyncFunction';
+export { isTokenExpired } from './isTokenExpired';
